+++ services-turf/src/main/java/com/mapbox/turf/TurfMeasurement.java
@@ -127,13 +127,18 @@ public final class TurfMeasurement {
    */
   public static double distance(@NonNull Point point1, @NonNull Point point2,
                                 @NonNull @TurfConstants.TurfUnitCriteria String units) {
-    double difLat = degreesToRadians((point2.latitude() - point1.latitude()));
-    double difLon = degreesToRadians((point2.longitude() - point1.longitude()));
-    double lat1 = degreesToRadians(point1.latitude());
-    double lat2 = degreesToRadians(point2.latitude());
+    double p1Latitude = point1.latitude();
+    double p1Longitude = point1.longitude();
+    double p2Latitude = point2.latitude();
+    double p2Longitude = point2.longitude();
+    double difLat = degreesToRadians(p2Latitude - p1Latitude);
+    double difLon = degreesToRadians(p2Longitude - p1Longitude);
+    double lat1 = degreesToRadians(p1Latitude);
+    double lat2 = degreesToRadians(p2Latitude);
+    double a = Math.sin(difLat / 2);
+    double b = Math.sin(difLon / 2);
 
-    double value = Math.pow(Math.sin(difLat / 2), 2)
-      + Math.pow(Math.sin(difLon / 2), 2) * Math.cos(lat1) * Math.cos(lat2);
+    double value = a * a + b * b * Math.cos(lat1) * Math.cos(lat2);
 
     return TurfConversion.radiansToLength(
       2 * Math.atan2(Math.sqrt(value), Math.sqrt(1 - value)), units);